added combinereducer move root reducer logic to reduxStorage, added radioButtonIllustrationResucer . fix: renamed Illustration component
