+++ src/views/TestArtDirectorAdmin/Components/CreateTaskContainer/Components/Illustrations/IllustrationGrig/IllustrationColumns/Illustration/Illustration.module.scss
